# api/docs/swagger.yaml
definitions:
  entity.CreateCategoryRequest:
    properties:
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.CreateCategoryResponse:
    properties:
      guid:
        type: string
    type: object
  entity.CreateCourierResponse:
    properties:
      guid:
        type: string
    type: object
...
  entity.DeleteCategoryRequest:
    properties:
      guid:
        type: string
    type: object
  entity.DeleteCourierRequest:
    properties:
      guid:
        type: string
    type: object
...
  entity.GetCategoryResponse:
    properties:
      created_at:
        type: string
      guid:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.GetCourierResponse:
    properties:
      created_at:
        type: string
      guid:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  entity.GetOrderResponse:
    properties:
      address:
        type: string
      branch_address:
        type: string
      branch_id:
        type: string
      branch_name:
        type: string
      comission:
        type: number
      comment:
        type: string
      courier_id:
        type: string
      created_at:
        type: string
      delivery_price:
        type: number
      delivery_time:
        type: number
      delivery_type:
        type: string
      guid:
        type: string
      image:
        type: string
      merchant_name:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProducts'
        type: array
      status:
        type: string
      user_id:
        type: string
    type: object
...
  entity.ListCategoryResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/entity.GetCategoryResponse'
        type: array
    type: object
  entity.ListCourierOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/entity.GetOrderResponse'
        type: array
    type: object
  entity.ListCourierResponse:
    properties:
      couriers:
        items:
          $ref: '#/definitions/entity.GetCourierResponse'
        type: array
    type: object
...
  entity.LoginCourierRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  entity.LoginCourierResponse:
    properties:
      exist:
        type: boolean
      guid:
        type: string
      name:
        type: string
    type: object
...
  entity.Option:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
...
  entity.OrderProducts:
    properties:
      category_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.ResponseOK:
    properties:
      message:
        type: object
    type: object
  entity.UpdateCategoryRequest:
    properties:
      guid:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.UpdateCategoryResponse:
    properties:
      guid:
        type: string
    type: object
  entity.UpdateCourierRequest:
    properties:
      guid:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  entity.UpdateCourierResponse:
    properties:
      guid:
        type: string
    type: object
...
  models.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.User:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/category/create:
    post:
      consumes:
      - application/json
      description: API that create category
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateCategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create category
      tags:
      - category
  /v1/category/delete:
    delete:
      consumes:
      - application/json
      description: API that delete category
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete category
      tags:
      - category
  /v1/category/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get category
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetCategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get category
      tags:
      - category
  /v1/category/list:
    get:
      consumes:
      - application/json
      description: API list Categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListCategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List Categories
      tags:
      - category
  /v1/category/update:
    put:
      consumes:
      - application/json
      description: API that update category
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateCategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update category
      tags:
      - category
  /v1/courier/{guid}/list-orders:
    get:
      consumes:
      - application/json
      description: API list courier orders
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListCourierOrdersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List courier orders
      tags:
      - courier
  /v1/courier/create:
    post:
      consumes:
      - application/json
      description: API that create courier
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateCourierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create courier
      tags:
      - courier
  /v1/courier/delete:
    delete:
      consumes:
      - application/json
      description: API that delete courier
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete courier
      tags:
      - courier
  /v1/courier/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get courier
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetCourierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get courier
      tags:
      - courier
  /v1/courier/list:
    get:
      consumes:
      - application/json
      description: API list couriers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListCourierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List couriers
      tags:
      - courier
  /v1/courier/login:
    post:
      consumes:
      - application/json
      description: Login courier
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginCourierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Login courier
      tags:
      - courier
  /v1/courier/update:
    put:
      consumes:
      - application/json
      description: API that update courier
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateCourierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update courier
      tags:
      - courier
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"